Add logging convenience functions
It's much nicer to have proper CBA-style logging with a category and
timestamp.  The functions make an effort to find a parent script if the first
argument is nil, but there is only so much they can do. Might want to add a
warning if the it is nil and _fnc_scriptName==_fnc_scriptNameParent, e.g. when
the logging function is called from the mission's init.sqf.