Add Murk to base PA Framework
Murk used to be only in the "extras" part of the framework and sometimes,
mission makers would forget to update `description.ext` or get the braces
slightly wrong when adding it. Always shipping it is the easier solution, and
does niether bloat the mission too much, nor does it cause any performance hit
when unused.